fix: Remove Windows => UNIX variable conversion
Before, Windows-style env variables (%test%) would be converted to UNIX-style format ($test) in UNIX
systems. That was an undocumented feature. I've removed that to simplify code a bit. Of course, UNIX
to Windows format still works, as that's the intended use of cross-env.

BREAKING CHANGE: %windows_style% env variables will no longer be converted to $unix_style in UNIX
machines. To fix it, use always the UNIX syntax, cross-env will change the format in Windows
machines as needed